tests are passing

+++ redis_glide.h
@@ -162,7 +162,7 @@ int execute_unlink_command(const void *glide_client, zval *keys, int keys_count,
 /* New commands added in redis_glide_commands.c */
 int execute_setOption_command(const void *glide_client, zend_long option, zval *value);
 int execute_mset_command(const void *glide_client, zval *arr);
-int execute_msetnx_command(const void *glide_client, zval *arr, int *output_value);
+int execute_msetnx_command(const void *glide_client, zval *arr, long *output_value);
 int execute_brpoplpush_command(const void *glide_client, const char *src, size_t src_len, const char *dst, size_t dst_len, zend_long timeout, char **result, size_t *result_len);
 
 /* Object operations */
